revert clock_t and size_t changes.

+++ sys/arch/amd64/include/ansi.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: ansi.h,v 1.6.18.3 2008/11/01 21:22:25 christos Exp $	*/
+/*	$NetBSD: ansi.h,v 1.6.18.4 2009/01/04 02:19:45 christos Exp $	*/
 
 /*-
  * Copyright (c) 1990, 1993
@@ -49,10 +49,10 @@
  *	#undef	_BSD_SIZE_T_
  *	#endif
  */
-#define	_BSD_CLOCK_T_		unsigned long	/* clock() */
+#define	_BSD_CLOCK_T_		unsigned int	/* clock() */
 #define	_BSD_PTRDIFF_T_		int		/* ptr1 - ptr2 */
-#define	_BSD_SIZE_T_		unsigned int	/* sizeof() */
-#define	_BSD_SSIZE_T_		int		/* byte count or error */
+#define	_BSD_SIZE_T_		unsigned long	/* sizeof() */
+#define	_BSD_SSIZE_T_		long		/* byte count or error */
 #define	_BSD_TIME_T_		__int64_t	/* time() */
 #ifdef __GNUC__
 #define	_BSD_VA_LIST_		__builtin_va_list /* GCC built-in type */
